updated new item data
