+++ forge-1.12.2-14.23.5.2838-mdk/src/main/resources/assets/minebikemod/lang/en_us.lang
@@ -1,7 +1,8 @@
 item.game_fish_rod.name=Game_fish_rod
-item.lettuce.name=Lettuce
+item.lettuce.name=Lettuce Leaf
 item.hamburger.name=Hamburger
 item.hamburgerbun.name=Hamburger Bun
 item.sandwich.name=Sandwich
 item.sandwichbread.name=Sandwich Bread
+item.lettucehead.name=Lettuce Head
 game.inprogress=Game Currently Running
